% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{buildSignatureMatrix}
\alias{buildSignatureMatrix}
\title{Build Signature Matrix}
\usage{
buildSignatureMatrix(scdata, markers, diff.cutoff = 0.5, pval.cutoff = 0.01)
}
\arguments{
\item{scdata}{single-cell RNA-seq count matrix}

\item{markers}{list of named data.frames with markers genes for each cluster}

\item{diff.cutoff}{minimum log2FC value}

\item{pval.cutoff}{minimum FDR value}
}
\value{
signature for to be used for \code{solveDampenedWLS}
}
\description{
build signature matrix using genes identified e.g. by DEAnalysis()
}
