% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSignatureMatrixUsingSeurat.R
\name{buildSignatureMatrixUsingSeurat}
\alias{buildSignatureMatrixUsingSeurat}
\title{Signature Matrix Using Seurat}
\usage{
buildSignatureMatrixUsingSeurat(
  scdata,
  id,
  path,
  diff.cutoff = 0.5,
  pval.cutoff = 0.01
)
}
\arguments{
\item{symbol}{}
}
\value{
Signature Matrix
}
\description{
This function builds a signature matrix using genes identified
  by the DEAnalysis() function.
}
\examples{

}
