% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataBulk.R
\docType{data}
\name{dataBulk}
\alias{dataBulk}
\title{Bulk data}
\format{
Data
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/28467820/}
}
\usage{
dataBulk
}
\description{
Bulk data for WT1 (control condition #1). More information on the data
can be found here: https://www.nature.com/articles/s41467-019-10802-z
}
\keyword{datasets}
