% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSC_1.R
\docType{data}
\name{dataSC_1}
\alias{dataSC_1}
\title{Single cell data}
\format{
Data
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/28467820/}
}
\usage{
dataSC_1
}
\description{
The "first half" of the single cell bulk data. To be combined with
dataSC_2 in the examples.
}
\keyword{datasets}
