% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSC_2.R
\docType{data}
\name{dataSC_2}
\alias{dataSC_2}
\title{Single cell data}
\format{
Data
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/28467820/}
}
\usage{
dataSC_2
}
\description{
The "second half" of the single cell bulk data. To be combined with
dataSC_1 in the examples.
}
\keyword{datasets}
