% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSC_3.R
\docType{data}
\name{dataSC_3}
\alias{dataSC_3}
\title{Single cell data}
\format{
Data
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/28467820/}
}
\usage{
dataSC_3
}
\description{
The "second half" of the single cell bulk data. 3 rows of dataSC_1
in the examples (full file is in inst/extdata) to be combined with dataSC_2
for the full dataset.
}
\keyword{datasets}
