% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DEAnalysis}
\alias{DEAnalysis}
\title{Get Seurat Markers}
\usage{
DEAnalysis(scdata, id, ...)
}
\arguments{
\item{scdata}{single-cell RNA-seq count matrix}

\item{id}{cluster assignment}

\item{...}{additional arguments to \code{Seurat::FindAllMarkers}}
}
\value{
list of data.frames with markers for each cluster
}
\description{
perform DE analysis using Seurat
}
