% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEAnalysis_Ident.R
\name{DEAnalysis_Ident}
\alias{DEAnalysis_Ident}
\title{Differential Expression with Idents (Seurat)}
\usage{
DEAnalysis_Ident(scdata, id, path)
}
\arguments{
\item{symbol}{}
}
\value{
Differential Expression Values
}
\description{
This function calculates the differential expression values
along with identifying the Idents (through Seurat). The output is
saved in an RDS file.
}
\examples{

}
