% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEAnalysisSeuratIdents.R
\name{DEAnalysisSeuratIdents}
\alias{DEAnalysisSeuratIdents}
\title{Differential Expression with Idents (Seurat)}
\usage{
DEAnalysisSeuratIdents(scdata, id, path)
}
\arguments{
\item{scdata}{The gene expression datafarme}

\item{id}{The unique identities within the data}

\item{path}{The path for the RData results}
}
\value{
An RData file with the differential expression analysis results
for each unique id.
}
\description{
This function calculates the differential expression values
along with identifying the Idents (through Seurat). The output is
saved in an RData file for each unique identity (id).
}
\examples{
\donttest{
download.file(
"https://github.com/sistia01/DWLS/raw/main/inst/extdata/dataSC.RData",
"dataSC.RData")
load("dataSC.RData")

#data('dataBulk', package = "DWLS")
download.file(
"https://github.com/sistia01/DWLS/raw/main/inst/extdata/dataBulk.RData",
"dataBulk.RData")
load("dataBulk.RData")

#data('labels', package = "DWLS")
download.file(
"https://github.com/sistia01/DWLS/raw/main/inst/extdata/labels.RData",
"labels.RData")
load("labels.RData")

#data('trueLabels', package = "DWLS")
download.file(
"https://github.com/sistia01/DWLS/raw/main/inst/extdata/trueLabels.RData",
"trueLabels.RData")
load("trueLabels.RData")

#Old Method
#load("data/dataSC_3.RData")
#load("data/trueLabels.RData")
#load("data/dataBulk.RData") #read in bulk data for WT1 (control condition #1)
#load("data/labels.RData") #read in single-cell labels from clustering

labels<-trueLabels
newcat<-c("NonCycISC","CycISC","TA","Ent","PreEnt","Goblet","Paneth",
"Tuft","EE")
for (i in 1:length(newcat)){
  labels[which(labels==(i-1))]<-newcat[i]
  }
#Run deconvolution
Seurat_test2 <- DEAnalysisSeuratIdents(dataSC, labels, "results")
}
}
