% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\docType{data}
\name{labels}
\alias{labels}
\title{Single-cell data cell-type labels}
\format{
A value (single-cell label)
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/28467820/}
}
\usage{
labels
}
\description{
These single cell labels were obtained from clustering. There are 9 unique
labels. "CycISC", "NonCycISC", "TA", "Goblet", "Paneth", "Ent", "PreEnt",
"Tuft", "EE"
}
\keyword{datasets}
