% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mean.in.log2space.R
\name{Mean.in.log2space}
\alias{Mean.in.log2space}
\title{Mean.in.log2space}
\usage{
Mean.in.log2space(x, pseudo.count)
}
\arguments{
\item{x}{Data}

\item{pseudo.count}{A pseudocount value}
}
\value{
Values
}
\description{
Applies the log2 to the mean of ((2^x - pseudo count) + pseudo
count).
}
\examples{
\donttest{
#dataBulk
#url <- "https://github.com/sistia01/DWLS/raw/main/inst/extdata/dataBulk.RData"
#dest <- "data/dataBulk.RData"
#download.file(url, dest)
#load("data/dataBulk.RData")
load(system.file("extdata", "dataBulk.RData", package = "DWLS"))

Mean.in.log2space(dataBulk, 0.1)
}

}
