% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sig.R
\docType{data}
\name{Sig}
\alias{Sig}
\title{Signature matrix}
\format{
Data
}
\usage{
Sig
}
\description{
This is an example of the output form buildSignatureMatrixUsingSeurat
function. A similar output is created using buildSignatureMatrixMAST.
}
\keyword{datasets}
