% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{solveDampenedWLS}
\alias{solveDampenedWLS}
\title{Solve Dampened WLS}
\usage{
solveDampenedWLS(S, B)
}
\arguments{
\item{S}{single-cell signature matrix}

\item{B}{bulk expression profile for one sample}
}
\value{
predicted cell-type proportions
}
\description{
solve using WLS with weights dampened by a certain dampening constant
}
