% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solveOLSInternal.R
\name{solveOLSInternal}
\alias{solveOLSInternal}
\title{solveOLSInternal}
\usage{
solveOLSInternal(S, B)
}
\arguments{
\item{S}{List output from trimData$sig (S)}

\item{B}{List output from trimData$bulk (B)}
}
\value{
Cell numbers
}
\description{
This function solves or the unknown parameters
using ordinary least squares (OLS) without printing the output.
It returns the cell numbers, not the proportions (see solveOLS).
}
\examples{

download.file(
"https://github.com/sistia01/DWLS/raw/main/inst/extdata/Sig.RData",
"Sig.RData")
load("Sig.RData")

#data('dataBulk', package = "DWLS")
download.file(
"https://github.com/sistia01/DWLS/raw/main/inst/extdata/dataBulk.RData",
"dataBulk.RData")
load("dataBulk.RData")

trimmed <- trimData(Sig, dataBulk)
S <- trimmed$sig
B <- trimmed$bulk
solveOLSInternal(S, B)

}
