% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{solveParallel}
\alias{solveParallel}
\title{Solve Parallel}
\usage{
solveParallel(
  S,
  Bmat,
  solver = solveDampenedWLS,
  ncores = min(ncol(Bmat), parallel::detectCores())
)
}
\arguments{
\item{S}{single-cell signature matrix}

\item{Bmat}{matrix of bulk expression signatures}

\item{solver}{function to use. Default is \code{solveDampenedWLS}.}

\item{ncores}{number of cores}
}
\value{
results
}
\description{
Run solver in parallel
}
