% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stat.log2.optimized}
\alias{stat.log2.optimized}
\title{A second version of the stat.log2 function that is optimized for higher processing speed}
\usage{
stat.log2.optimized(data.m, group.v, pseudo.count)
}
\description{
A second version of the stat.log2 function that is optimized for higher processing speed
}
