% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.log2.R
\name{stat.log2}
\alias{stat.log2}
\title{stat.log2}
\usage{
stat.log2(data.m, group.v, pseudo.count)
}
\arguments{
\item{data.m}{Data}

\item{group.v}{Groupings}

\item{pseudo.count}{A pseudocount value}
}
\value{
A dataframe of the log2 applied results
}
\description{
One of the functions required for the  differential expression
analysis using MAST (DEAnalysisMast()) function.
}
\examples{
pseudo.count = 0.1
dataSC <- cbind(dataSC_1, dataSC_2)
data.used.log2   <- log2(dataSC+pseudo.count)
colnames(data.used.log2)<-make.unique(colnames(data.used.log2))
diff.cutoff=0.5
for (i in unique(labels)){
  cells.symbol.list2 = colnames(data.used.log2)[which(labels==i)]
  cells.coord.list2 = match(cells.symbol.list2, colnames(data.used.log2))
  cells.symbol.list1 = colnames(data.used.log2)[which(labels != i)]
  cells.coord.list1 = match(cells.symbol.list1, colnames(data.used.log2))
  data.used.log2.ordered = cbind(data.used.log2[,cells.coord.list1],
                                        data.used.log2[,cells.coord.list2])}
group.v <- c(rep(0,length(cells.coord.list1)),
                                         rep(1, length(cells.coord.list2)))

}
