% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{trimData}
\alias{trimData}
\title{Trim Signature}
\usage{
trimData(Signature, bulkData)
}
\arguments{
\item{Signature}{single-cell signature from \code{buildSignatureMatrix}}

\item{bulkData}{bulk expression matrix (genes x samples)}
}
\value{
list with single-cell and bulk matrices trimmed to common genes
}
\description{
Trim bulk and single-cell data to contain the same genes
}
