% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trueLabels.R
\docType{data}
\name{trueLabels}
\alias{trueLabels}
\title{Single cell data labels}
\format{
Data
}
\source{
\url{https://github.com/dtsoucas/DWLS/tree/master/ISC/data}
}
\usage{
trueLabels
}
\description{
The "true labels" for the example single-cell labels from clustering.
}
\keyword{datasets}
