% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v.auc.R
\name{v.auc}
\alias{v.auc}
\title{v.auc}
\usage{
v.auc(data.v, group.v)
}
\arguments{
\item{data.v}{Data}

\item{group.v}{Data subdivision}
}
\value{
Matrix of standardized output of AUC calculation
}
\description{
Uses the prediction() function in order to create standardized
output from the data in order to perform an AUC calculation. The calculation
results are rounded to the third decimal place. This function serves mainly
to support the DWLS function.
}
\examples{

\dontrun{
m.auc=function(data.m,group.v)
{ AUC=apply(data.m, 1, function(x) v.auc(x,group.v))
AUC[is.na(AUC)]=0.5
return(AUC)}
}

}
