% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{buildSignatureMatrix}
\alias{buildSignatureMatrix}
\title{Build Signature Matrix}
\usage{
buildSignatureMatrix(
  scdata,
  id,
  markers,
  diff.cutoff = 0.5,
  pval.cutoff = 0.01
)
}
\arguments{
\item{scdata}{single-cell RNA-seq count matrix}

\item{id}{cluster assignment}

\item{markers}{list of data.frames with markers genes for each cluster with columns
'avg_log2FC' and 'p_val_adj' and \code{row.names} as gene symbols.}

\item{diff.cutoff}{minimum 'avg_log2FC' value}

\item{pval.cutoff}{minimum 'p_val_adj' value}
}
\value{
signature for to be used for \code{solveDampenedWLS}
}
\description{
build signature matrix using genes identified e.g. by DEAnalysis()
}
