% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSignatureMatrixMAST.R
\name{buildSignatureMatrixMAST}
\alias{buildSignatureMatrixMAST}
\title{Signature Matrix Using MAST}
\usage{
buildSignatureMatrixMAST(
  scdata,
  id,
  path,
  diff.cutoff = 0.5,
  pval.cutoff = 0.01,
  f = 200
)
}
\arguments{
\item{scdata}{The data}

\item{id}{The identities of the genes}

\item{path}{The path to the file results}

\item{diff.cutoff}{This is automatically set to 0.5}

\item{pval.cutoff}{This is automatically set to 0.01}

\item{f}{The maximum number of genes (when creating the signature matrix,
need to reduce number of genes, between 50:f number of significant genes are
chosen). If not set, this number is automatically set to 200.}
}
\value{
Signature Matrix built using the MAST algorithm
}
\description{
This function builds a signature matrix using genes identified
by the DEAnalysisMAST() function.
}
\examples{
\donttest{
download.file("https://github.com/sistia01/DWLS/raw/main/inst/extdata/dataSC.RData", "dataSC.RData")
load("dataSC.RData")
data('dataBulk', package = "DWLS")
data('labels', package = "DWLS")
data('trueLabels', package = "DWLS")

#Old Method
#load("data/dataSC_3.RData")
#load("data/trueLabels.RData")
#load("data/dataBulk.RData") #read in bulk data for WT1 (control condition #1)
#load("data/labels.RData") #read in single-cell labels from clustering

labels<-trueLabels

newcat<-c("NonCycISC","CycISC","TA","Ent","PreEnt","Goblet","Paneth","Tuft","EE")
for (i in 1:length(newcat)){
  labels[which(labels==(i-1))]<-newcat[i]
  }

Signature<-buildSignatureMatrixMAST(dataSC,labels,"inst/extdata/results", diff.cutoff=0.5,pval.cutoff=0.01)
}
}
