% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSC_3.R
\docType{data}
\name{dataSC_3}
\alias{dataSC_3}
\title{Single cell data}
\format{
Data
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/28467820/}
}
\usage{
dataSC_3
}
\description{
A subset of the dataSC dataset in inst/extdata/
}
\keyword{datasets}
