% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEAnalysisMAST.R
\name{DEAnalysisMAST}
\alias{DEAnalysisMAST}
\title{DEAnalysisMAST}
\usage{
DEAnalysisMAST(scdata, id, path)
}
\arguments{
\item{scdata}{The gene expression datafarme}

\item{id}{The unique identities within the data}

\item{path}{The path for the RData results}
}
\value{
matrix
The resulting matrix is a gene by cell-type signature matrix.
The cell-type signature matrix is constructed using a representative
single-cell data set, such that all cell types expected in the bulk
data are also represented in the single-cell data (the converse need not be
true). The single-cell data is first clustered to reveal
its constituent cell types.The function will return 3 different files:
an RData file, an rds file, and a csv file.
}
\description{
Perform DE analysis using MAST.
Dampened weighted least squares (DLWS) is an estimation
method for gene expression deconvolution, in which the cell-type
composition of a bulk RNA-seq data set is computationally inferred.
This method corrects common biases towards cell types that are
characterized by highly expressed genes and/or are highly prevalent,
to provide accurate detection across diverse cell types. To begin,
the user must input a bulk RNA-seq data set, along with a labeled
representative single-cell RNA-seq data set that will serve to generate
cell-type-specific gene expression profiles. Ideally, the single-cell
data set will contain cells from all cell types that may be found in the
bulk data. DWLS will return the cell-type composition of the bulk data.
First, solve OLS then use the solution to find a starting point for the
weights. Next, the dampened weighted least squares is performed. The weights
are iterated until convergence then the dampening constant for weights is
found using cross-validation (with decreasing step size for convergence).

DWLS captures ISC composition changes across conditions.
One of the most important applications of deconvolution methods is in
the identification of cell-type composition variations across conditions.

Note: The function uses solveDampenedWLSj() and findDampeningConstant().
}
\examples{
\dontrun{
load("data/dataSC_1.RData")
load("data/dataSC_1.RData")
load("data/trueLabels.RData")
load("data/dataBulk.RData") #read in bulk data for WT1 (control condition #1)
load("data/labels.RData") #read in single-cell labels from clustering
labels<-trueLabels
dataSC <- cbind(dataSC_1, dataSC_2)
newcat<-c("NonCycISC","CycISC","TA","Ent","PreEnt","Goblet","Paneth","Tuft","EE")
for (i in 1:length(newcat)){
  labels[which(labels==(i-1))]<-newcat[i]
  }
#Run deconvolution
Mast_test <- DEAnalysisMAST(dataSC, labels, "results")
}

}
