% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEAnalysisSeurat.R
\name{DEAnalysisSeurat}
\alias{DEAnalysisSeurat}
\title{Differential Expression without Idents (Seurat)}
\usage{
DEAnalysisSeurat(scdata, id, path)
}
\arguments{
\item{scdata}{The gene expression datafarme}

\item{id}{The unique identities within the data}

\item{path}{The path for the RData results}
}
\value{
An RData file with the differential expression analysis results
for each unique id.
}
\description{
This function calculates the differential expression values
along with identifying the Idents (through Seurat). The output is
saved in an RData file for each unique identity (id).
}
\examples{
\dontrun{
load("data/dataSC_1.RData")
load("data/dataSC_2.RData")
dataSC <- cbind(dataSC_1, dataSC_2)
load("data/trueLabels.RData")
load("data/dataBulk.RData") #read in bulk data for WT1 (control condition #1)
load("data/labels.RData") #read in single-cell labels from clustering
labels<-trueLabels
newcat<-c("NonCycISC","CycISC","TA","Ent","PreEnt","Goblet","Paneth","Tuft","EE")
for (i in 1:length(newcat)){
  labels[which(labels==(i-1))]<-newcat[i]
  }
#Run deconvolution
Seurat_test2 <- DEAnalysisSeurat(dataSC, labels, "results")
}

}
