% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solveDampenedWLSj.R
\name{solveDampenedWLSj}
\alias{solveDampenedWLSj}
\title{solveDampenedWLSj}
\usage{
solveDampenedWLSj(S, B, goldStandard, j)
}
\arguments{
\item{S}{List output from trimData$sig (S)}

\item{B}{List output from trimData$bulk (B)}

\item{goldStandard}{Starting point for the weights, this can be determined
using solveOLSInternal(S,B)}

\item{j}{The dampening constant, this can be determined using
findDampeningConstant(S,B,goldStandard)}
}
\value{
value (Dampened weighted least squares estimation values)
}
\description{
Solve dampened weighted least squares given a dampening constant.

Note: The function uses solveDampenedWLS() and findDampeningConstant().
}
\examples{

#load("data/Sig.RData")
#load("data/dataBulk.RData")
data('Sig', package = "DWLS")
data('dataBulk', package = "DWLS")

trimmed <- trimData(Sig, dataBulk)
S <- trimmed$sig
B <- trimmed$bulk
solution <- solveOLSInternal(S,B)
j <- findDampeningConstant(S,B,solution)
goldStandard <- solveOLSInternal(S,B)
solveDampenedWLSj(S,B,goldStandard,j)

}
