% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solveOLS.R
\name{solveOLS}
\alias{solveOLS}
\title{solveOLS}
\usage{
solveOLS(S, B)
}
\arguments{
\item{S}{List output from trimData$sig (S)}

\item{B}{List output from trimData$bulk (B)}
}
\value{
Cell-type proportion
}
\description{
This function solves or the unknown parameters
using ordinary least squares (OLS). It is constrained such
that cell type numbers are greater than 0.
}
\examples{
trimData(Sig, dataBulk)
B <- test$bulk
solveOLS(S, B)

}
