% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solveOLS.R
\name{solveOLS}
\alias{solveOLS}
\title{solveOLS}
\usage{
solveOLS(S, B)
}
\arguments{
\item{bulkdata}{}
}
\description{
This function solves using OLS. It is constrained such that cell type numbers are greater than 0
}
\examples{

}
