% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solveOLS.R
\name{solveOLS}
\alias{solveOLS}
\title{solveOLS}
\usage{
solveOLS(S, B)
}
\arguments{
\item{S}{List output from trimData$sig (S)}

\item{B}{List output from trimData$bulk (B)}
}
\value{
Cell-type proportion
}
\description{
This function solves or the unknown parameters
using ordinary least squares (OLS). It is constrained such
that cell type numbers are greater than 0.
}
\examples{

download.file(
"https://github.com/sistia01/DWLS/raw/main/inst/extdata/Sig.RData",
"Sig.RData")
load("Sig.RData")

#data('dataBulk', package = "DWLS")
download.file(
"https://github.com/sistia01/DWLS/raw/main/inst/extdata/dataBulk.RData",
"dataBulk.RData")
load("dataBulk.RData")

trimmed <- trimData(Sig, dataBulk)
S <- trimmed$sig
B <- trimmed$bulk
solveOLS(S, B)

}
