% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimData.R
\name{trimData}
\alias{trimData}
\title{trimData}
\usage{
trimData(Signature, bulkData)
}
\arguments{
\item{bulkdata}{}
}
\description{
This function trims bulk and single-cell data to contain the same genes
}
\examples{
trimData(Signature, bulkData)

}
