% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimData.R
\name{trimData}
\alias{trimData}
\title{trimData}
\usage{
trimData(Signature_Matrix, bulkdata)
}
\arguments{
\item{Signature_Matrix}{A single-cell signature matrix}

\item{bulkdata}{A bulk dataset}
}
\value{
A list of trimmed bulk and single-cell data.
}
\description{
This function trims bulk and single-cell data to
contain the same genes. The result is a list of the intersecting genes
within the two datasets.
}
\examples{
 \donttest{
data('Sig', package = "DWLS")
data('dataBulk', package = "DWLS")

trimData(Signature_Matrix = Sig, bulkdata = dataBulk)
}
}
